# doc/gen-res-by-zupportL5/static/public/ZupportL5-api-swagger-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ZupportL5 API Documentation",
    "description": "Comprehensive OpenAPI-compliant JSON documentation for ZupportL5-API",
    "version": "1.0.0",
    "contact": {
      "email": "arcbrth@gmail.com",
      "name": "Development"
    },
    "license": {
      "url": "https://github.com/abritoh/zupportl5-api",
      "name": "Apache 2.0"
    },
    "origin": "This documentation was generated at 2025-02-14T16:17:34Z using ZupportL5 json-api-processor"
  },
  "paths": {
    "/api/runbooks/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          },
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "runbook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "deleteById",
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/runbooks/": {
      "post": {
        "summary": "create",
        "operationId": "create",
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "runbook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getAll",
        "operationId": "getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/api/priorities/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "priority"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "delete",
        "operationId": "delete",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/priorities/": {
      "post": {
        "summary": "create",
        "operationId": "create",
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "priority"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getAll",
        "operationId": "getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/api/resolution-steps/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "resolutionstep"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "deleteById",
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/resolution-steps/": {
      "post": {
        "summary": "save",
        "operationId": "save",
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "resolutionstep"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getAll",
        "operationId": "getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/api/playbooks/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          },
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "playbook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "deleteById",
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "long"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/api/status/{id}": {
      "put": {
        "summary": "update",
        "operationId": "update",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "item",
            "in": "body",
            "required": true,
            "schema": {
              "type": "status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "summary": "delete",
        "operationId": "delete",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getById",
        "operationId": "getById",
        "parameters": [
          {
            "name": "",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/status/": {
      "post": {
        "summary": "create",
        "operationId": "create",
        "parameters": [
          {
            "name": "runbook",
            "in": "body",
            "required": true,
            "schema": {
              "type": "status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getAll",
        "operationId": "getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  }
}